[Bronze II] Title: 숫자, Time: 24 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -4,18 +4,18 @@
 using namespace std;
 
 int main() {
-	int A, B;
-	int m, M;
+	long long A, B;
+	long long m, M;
 
 	cin >> A >> B;
 
 	M = max(A, B);
 	m = min(A, B);
 
-	if (M == m) printf("%d\n", M - m);
-	else printf("%d\n", M - m - 1);
+	if (M == m) printf("%lld\n", M - m);
+	else printf("%lld\n", M - m - 1);
 
-	for (int i = m + 1; i < M; i++) printf("%d ", i);
+	for (long long i = m + 1; i < M; i++) printf("%lld ", i);
 
 	return 0;
 }
